requirements.txt
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.2
Flask-Login==0.6.3
flask-mongoengine-3==1.1.0
Flask-WTF==1.2.2
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mongoengine==0.29.1
packaging==25.0
pymongo==4.14.1
python-dotenv==1.1.1
regex==2025.7.34
requests==2.32.5
urllib3==2.5.0
watchdog==6.0.0
Werkzeug==3.1.3
WTForms==3.2.1
